+++ pbge/randmaps/rooms.py
@@ -24,6 +24,7 @@ class Room(object):
     ON_THE_EDGE = False
     MIN_RANDOM_SIZE = 7
     MAX_RANDOM_SIZE = 12
+    EXPAND_AMOUNT = 5
 
     def __init__(self, width=None, height=None, tags=(), anchor=None, parent=None, archi=None, decorate=None, **kwargs):
         self.width = width or random.randint(self.MIN_RANDOM_SIZE, self.MAX_RANDOM_SIZE)
@@ -87,10 +88,25 @@ class Room(object):
             if isinstance(r, Room):
                 r.step_seven(gb, archi)
 
+    #def all_rooms(self):
+    #    yield self
+    #    for r in self.contents:
+    #        if isinstance(r, Room):
+    #            for rr in r.all_rooms():
+    #                yield rr
+
+    def _expand(self):
+        # This room isn't big enough. Make it bigger, then raise an error.
+        self.width += self.EXPAND_AMOUNT
+        self.height += self.EXPAND_AMOUNT
+        raise RoomError("ROOM ERROR: {}:{} isn't big enough".format(str(self), str(self.__class__)), self)
+
     def find_spot_for_room(self, closed_area, myroom):
         myrect = pygame.Rect(0, 0, myroom.width, myroom.height)
         count = 0
-        while (count < 1000) and not myroom.area:
+        if self.area.width <= myroom.width or self.area.height <= myroom.height:
+            self._expand()
+        while (count < 100) and not myroom.area:
             myrect.x = random.choice(list(range(self.area.x, self.area.x + self.area.width - myroom.width)))
             myrect.y = random.choice(list(range(self.area.y, self.area.y + self.area.height - myroom.height)))
             if self.ON_THE_EDGE and count < 500:
@@ -110,8 +126,7 @@ class Room(object):
                 if random.randint(1, 3) == 3 and myrect.height > myroom.MIN_RANDOM_SIZE:
                     myrect.height -= 1
             count += 1
-        if not myroom.area:
-            raise RoomError("ROOM ERROR: {}:{} cannot place {}".format(str(self), str(self.__class__), str(myroom)))
+        return(bool(myroom.area))
 
     def arrange_contents(self, gb):
         # Step Two: Arrange subcomponents within this area.
@@ -133,25 +148,26 @@ class Room(object):
         unplaced_rooms = [r for r in self.contents if hasattr(r, "area") and not r.area]
         tries = 100
         while tries > 0:
-            try:
-                for r in unplaced_rooms:
-                    ok = self.find_spot_for_room(closed_area, r)
+            ok = True
+            for r in unplaced_rooms:
+                ok = ok and self.find_spot_for_room(closed_area, r)
+
+            if ok:
                 break
-            except RoomError:
+            else:
                 for r in unplaced_rooms:
                     if r.area in closed_area:
                         closed_area.remove(r.area)
                     r.area = None
                     tries -= 1
-                    if (tries < 50 or random.randint(1, 3) == 3) and r.width > r.MIN_RANDOM_SIZE:
-                        r.width -= 1
-                    if (tries < 50 or random.randint(1, 3) == 3) and r.height > r.MIN_RANDOM_SIZE:
-                        r.height -= 1
-                    if tries < 10:
-                        print("Warning: Running out of tries!")
+                    #if (tries < 50 or random.randint(1, 3) == 3) and r.width > r.MIN_RANDOM_SIZE:
+                    #    r.width -= 1
+                    #if (tries < 50 or random.randint(1, 3) == 3) and r.height > r.MIN_RANDOM_SIZE:
+                    #    r.height -= 1
+                    #if tries < 10:
+                    #    print("Warning: Running out of tries!")
         if tries < 1:
-            raise RoomError(
-                "ROOM ERROR: {}:{} cannot place all requested rooms.".format(str(self), str(self.__class__)))
+            self._expand()
 
     def connect_contents(self, gb, archi):
         # Step Three: Connect all rooms in contents, making trails on map.